chore: create `.gitattributes` for consistent EOL handling (#144)
This pull request includes changes to ensure consistent end-of-line
(EOL) settings across all text files and updates to the synchronization
configuration for the repository. The most important changes are:

EOL consistency:

*
[`.gitattributes`](diffhunk://#diff-618cd5b83d62060ba3d027e314a21ceaf75d36067ff820db126642944145393eR1-R4):
Added configurations to ensure consistent EOL (LF) for all text files.

Synchronization configuration updates:

*
[`.github/sync-client.yml`](diffhunk://#diff-93bc202766315b6269beef308a6ad30ed3e86938ddbfa31b49e030f2263695f1R52-R53):
Added `.gitattributes` to the list of files to be synchronized to the
`./configs` directory.