docs: update `CONTRIBUTING.md` with guidelines on using the `node:` namespace (#146)
This pull request includes an update to the `CONTRIBUTING.md` file to
provide guidance on the usage of the `node:` namespace prefix for
built-in modules in Node.js. The most important change is the addition
of a new section explaining the recommended usage patterns for different
parts of the project.

Documentation updates:

*
[`CONTRIBUTING.md`](diffhunk://#diff-eca12c0a30e25b4b46522ebf89465a03ba72a03f540796c979137931d8f92055R48-R69):
Added a new section titled "Others" with a subsection on the `node:`
namespace. It explains that standard imports without the `node:` prefix
should be used for regular application code for broader compatibility,
while the `node:` prefix can be used in test files that use the
`node:test` runner.