chore: ready for introduction of website using `nextra` (#148)
This pull request includes several important changes to the
configuration files and the project structure. The most significant
changes involve updates to the `.vscode` settings, the addition of a new
workspace, and the creation of a `package.json` for the `website`
directory.

Updates to `.vscode` settings:

*
[`.github/sync-client.yml`](diffhunk://#diff-93bc202766315b6269beef308a6ad30ed3e86938ddbfa31b49e030f2263695f1L42-R42):
Changed the destination of the `.vscode/extensions.json` file from
`npm-extensions.json` to `web-extensions.json`.
*
[`.vscode/extensions.json`](diffhunk://#diff-c16655a98a3ee89a7636a59c59a72b0e93649e3a1e947327cfc43a1336b4e912R4):
Added `stylelint.vscode-stylelint` to the list of recommended
extensions.

Project structure changes:

*
[`package.json`](diffhunk://#diff-7ae45ad102eab3b6d7e7896acd08c427a9b25b346470d7bc6507b6481575d519L6-R7):
Added `website` to the list of workspaces.
*
[`website/package.json`](diffhunk://#diff-fae242fbf77a8a9d52625664bc36ea12316586f8a716b41137d897a2b7e3df76R1-R8):
Created a new `package.json` file for the `website` directory with basic
configuration, including a `start` script.