chore: bump macOS version references to `13` in CI workflow and documentation (#150)
This pull request includes updates to the macOS versions used in the
GitHub Actions workflows and the corresponding documentation. The most
important changes include updating the macOS version from 12 to 13 and
ensuring the documentation reflects these changes.

Updates to macOS versions:

*
[`.github/workflows/llvm-build-bump-pr.yml`](diffhunk://#diff-d608eb75b383338d789a8908995e47499214a7cc55a847771ad73800464b18b0L249-R249):
Changed the macOS version from 12 to 13 for x64 architecture in the
matrix configuration.

Documentation updates:

*
[`docs/01-introduction/05-supported.md`](diffhunk://#diff-664f5e525371f1677c342e11b06a53a625232a2dad8f12bbde8eeab27e037b09L54):
Removed references to macOS 12 and updated them to macOS 13 in the
supported platforms list.
*
[`docs/01-introduction/05-supported.md`](diffhunk://#diff-664f5e525371f1677c342e11b06a53a625232a2dad8f12bbde8eeab27e037b09L77-R76):
Updated the binary folder name table to reflect the change from macOS 12
to macOS 13 for `darwin-x64`.