docs: update `README.md` of `examples-clang-format` (#161)
This pull request includes updates to the
`examples/clang-format/README.md` file to improve clarity and provide
more comprehensive instructions for running the example. The most
important changes include renaming the example, restructuring the
instructions, and expanding the list of commands.

Improvements to documentation:

* Renamed the example from `clang-format` to `examples-clang-format` for
consistency.
* Restructured the instructions by adding a "Getting Started" section
and separating the methods for running the example into "Run from the
Root" and "Navigate to the Directory and Run" sections.
* Expanded the list of available commands to include all variations of
formatted and unformatted commands for both C and C++ languages.