ci: update labeler workflow to trigger on edited pull requests (#182)
This pull request includes a change to the GitHub Actions workflow for
labeling pull requests. The change ensures that the labeler action runs
not only when a pull request is opened but also when it is edited.

*
[`.github/workflows/labeler.yml`](diffhunk://#diff-09b72f3c9a3e4f00ab00cd7000b302db25f056075d8895bd91b3654d6e7e956bL5-R13):
Modified the `pull_request_target` event to include the `edited` type
and added a condition to run the labeler step when the pull request is
opened or its title is changed.