ci: add `concurrency` settings to `test-cross-platform.yml` test workflow (#183)
This pull request includes a change to the GitHub Actions workflow
configuration to manage concurrency for the `test-cross-platform.yml`
workflow. The change ensures that concurrent runs of the workflow are
managed more efficiently by grouping them and canceling in-progress runs
for non-main branches.

Concurrency management:

*
[`.github/workflows/test-cross-platform.yml`](diffhunk://#diff-290907d9a4bcf13181105b134bd1645da5a3ef27c54f148ad5823e6b77e88426R14-R17):
Added `concurrency` configuration to group workflow runs by workflow
name and reference, and to cancel in-progress runs for non-main
branches.