refactor(website): add Google Analytics integration to `.vitepress/config.mjs` (#184)
This pull request includes changes to add Google Analytics to the
website configuration and a minor documentation correction. The most
important changes include adding the Google Analytics ID and script to
the configuration file and fixing a typo in the documentation.

### Configuration updates:

*
[`website/.vitepress/config.mjs`](diffhunk://#diff-bd4ac01d4c35583e9c20414103304b0628a6be199987bd1084fdef61cd43cc71R28):
Added `GOOGLE_GA_ID` constant and included Google Analytics script tags
for tracking.
[[1]](diffhunk://#diff-bd4ac01d4c35583e9c20414103304b0628a6be199987bd1084fdef61cd43cc71R28)
[[2]](diffhunk://#diff-bd4ac01d4c35583e9c20414103304b0628a6be199987bd1084fdef61cd43cc71L70-L76)

### Documentation correction:

*
[`website/docs/get-started/configuration.md`](diffhunk://#diff-6a87b005e020bd2dfcf1e779a14c7ddc13bfc307a99b1b02b6c538e7f4da1b77L59-R59):
Corrected a typo from "an brief example" to "a brief example".