docs: create `q-and-a.md` (#185)
This pull request adds a new "Q & A" section to the documentation and
updates the website configuration to include a link to this new section.

Documentation updates:

* Added a new "Q & A" section to the documentation with questions and
answers about `clang-format-node`, including an explanation of the
differences between `clang-format` and `git-clang-format`.
(`website/docs/get-started/q-and-a.md`)

Website configuration updates:

* Updated the VitePress configuration to include a link to the new "Q &
A" section in the navigation menu. (`website/.vitepress/config.mjs`)