docs: create `surgical-formatting-with-git-clang-format.md` (#187)
This pull request adds a new documentation page and updates the site
configuration to include a link to this new page. The most important
changes include adding a new entry to the navigation menu and creating
detailed content for the new documentation page on surgical formatting
with `git-clang-format`.

Documentation updates:

*
[`website/.vitepress/config.mjs`](diffhunk://#diff-bd4ac01d4c35583e9c20414103304b0628a6be199987bd1084fdef61cd43cc71R257-R260):
Added a new entry "Surgical Formatting" to the navigation menu to link
to the new documentation page.
*
[`website/docs/further-reading/surgical-formatting-with-git-clang-format.md`](diffhunk://#diff-4eabe91a6c51648711792a2daa82f0a8d9efcf9d01642f6940cd47868407fd84R1-R105):
Created a new documentation page that explains how to use
`git-clang-format` for surgically formatting only the changes in a pull
request, including detailed steps and examples.