ci: update commit type of `llvm-build-bump-pr.yml` to `chore` from `build` (#189)
This pull request includes changes to update the commit message and
title format in the LLVM bump workflow and a minor documentation update.

Changes to workflow:

*
[`.github/workflows/llvm-build-bump-pr.yml`](diffhunk://#diff-d608eb75b383338d789a8908995e47499214a7cc55a847771ad73800464b18b0L429-R432):
Updated the commit message and title prefix from 'build(deps)' to
'chore(deps)' for consistency with other dependency updates.

Documentation update:

*
[`website/docs/get-started/introduction.md`](diffhunk://#diff-aeb0fea669d0b87f99bf331106ddc13e198f254aa732266a2435acf9a98f82aaL143-R143):
Added a colon at the end of the sentence to clarify the explanation of
the `git diff` command output.