chore: add `dev` and `start` scripts for every packages (#197)
This pull request introduces new build and development scripts for the
`clang-format` packages and updates the `package.json` files
accordingly. The most important changes include adding dev and start
scripts for the `clang-format-git`, `clang-format-git-python`, and
`clang-format-node` packages, and updating the main `package.json` to
include these new scripts.

### Updates to build and development scripts:

*
[`package.json`](diffhunk://#diff-7ae45ad102eab3b6d7e7896acd08c427a9b25b346470d7bc6507b6481575d519L16-R30):
Added new build and development scripts for `clang-format-git`,
`clang-format-git-python`, and `clang-format-node` packages.
*
[`packages/clang-format-git/package.json`](diffhunk://#diff-d85dcfb0934f579cff896723ab5eeb9a52668d2bb322338dddd93bc1241aa4f1L52-R54):
Added `dev` and `start` scripts to facilitate development and execution
of the `clang-format-git` package.
*
[`packages/clang-format-git-python/package.json`](diffhunk://#diff-fb0eb5dba5b2656b3fdea4322c2ed2a691973f2e9ea536ce20b36c347ab1267eL53-R55):
Added `dev` and `start` scripts to facilitate development and execution
of the `clang-format-git-python` package.
*
[`packages/clang-format-node/package.json`](diffhunk://#diff-6e04d643c57b851ab02ad78e41531387d2cc01caf6c4cfbaa536aadd37346a30L51-R53):
Added `dev` and `start` scripts to facilitate development and execution
of the `clang-format-node` package.