chore: add `vercel-ignore-command.sh` script and update `vercel.json` configuration (#205)
This pull request includes changes to add a new script for Vercel build
ignoring, update the Vercel configuration, and modify the sync client
configuration. The most important changes are:

### New Script Addition:
* Added a new script `vercel-ignore-command.sh` to ignore builds for
pull requests created by `dependabot[bot]` if the branch is not `main`.

### Vercel Configuration Updates:
* Updated `vercel.json` to include the new `ignoreCommand`, set the
build and install commands, specify the framework as `vitepress`, and
define the output directory.

### Sync Client Configuration Update:
* Updated `.github/sync-client.yml` to include the new script in the
sync process.