docs: correct GitHub PR list link in `SECURITY.md` (#210)
This pull request includes a minor update to the `SECURITY.md` file to
improve the clarity of the instructions for reviewing pull requests
related to LLVM builds.

Documentation update:

*
[`SECURITY.md`](diffhunk://#diff-f6ed156e4bf5c791680662464b94ea5d753f219ee816b385f67870e2c0d7d4c7L12-R12):
Updated the link to the Pull Request list on GitHub to use a more
specific query for filtering pull requests related to LLVM dependency
bumps.