chore(*): set `open-pull-requests-limit` to 3 in `dependabot.yml` configuration (#231)
This pull request includes changes to the `.github/dependabot.yml` file
to limit the number of open pull requests created by Dependabot. The
change sets a limit on the number of open pull requests to ensure that
the repository does not get overwhelmed with too many updates at once.

*
[`.github/dependabot.yml`](diffhunk://#diff-dd4fbda47e51f1e35defb9275a9cd9c212ecde0b870cba89ddaaae65c5f3cd28R30):
Added `open-pull-requests-limit: 3` to the `patterns` section to limit
the number of open pull requests to 3.
*
[`.github/dependabot.yml`](diffhunk://#diff-dd4fbda47e51f1e35defb9275a9cd9c212ecde0b870cba89ddaaae65c5f3cd28R50):
Added `open-pull-requests-limit: 3` to the `directory` section to limit
the number of open pull requests to 3.