ci(*): drop `bump.yml` and create new `version-monorepo.yml` workflow (#247)
This pull request includes several changes related to the configuration
and workflows of the repository. The most significant changes involve
the reorganization and updating of GitHub Actions workflows,
particularly the removal of the `bump.yml` workflow and the addition of
the `version-monorepo.yml` workflow.

### Workflow Updates and Reorganization:

* Removed the `bump.yml` workflow, which handled version bumps for
different types of releases.
* Added a new `version-monorepo.yml` workflow to handle version bumps
with more detailed configuration options, including semver and preid
selections. This workflow also drafts releases and creates pull
requests.

### Configuration File Updates:

* Updated `.editorconfig-checker.json` to exclude `lerna.json` from
checks.
* Updated `.github/sync-client.yml` to reflect changes in the workflows
being synchronized, including the removal of `bump.yml` and the addition
of `version-monorepo.yml`.
[[1]](diffhunk://#diff-93bc202766315b6269beef308a6ad30ed3e86938ddbfa31b49e030f2263695f1L12-L13)
[[2]](diffhunk://#diff-93bc202766315b6269beef308a6ad30ed3e86938ddbfa31b49e030f2263695f1R22-R23)