chore(website): conditionally enable bundle analysis in Codecov Vite plugin (#267)
This pull request includes changes to the
`website/.vitepress/config.mjs` and `website/package.json` files to
improve the configuration and dependency management for the project. The
most important changes include the addition of the `is-interactive`
package and adjustments to the Codecov plugin configuration to work
better in CI environments.

Dependency management and configuration improvements:

*
[`website/.vitepress/config.mjs`](diffhunk://#diff-bd4ac01d4c35583e9c20414103304b0628a6be199987bd1084fdef61cd43cc71R16):
Added import for the `is-interactive` package to the configuration file.
*
[`website/.vitepress/config.mjs`](diffhunk://#diff-bd4ac01d4c35583e9c20414103304b0628a6be199987bd1084fdef61cd43cc71L341-R342):
Modified the `enableBundleAnalysis` option for the Codecov plugin to use
the `isInteractive()` function, ensuring it only runs in CI
environments.
*
[`website/package.json`](diffhunk://#diff-fae242fbf77a8a9d52625664bc36ea12316586f8a716b41137d897a2b7e3df76R13):
Added `is-interactive` as a development dependency.