ci(*): create `release.yml` (#281)
This pull request introduces a new workflow configuration for handling
releases and updates the sync-client configuration to include this new
workflow. The most important changes are:

Workflow configuration updates:

*
[`.github/sync-client.yml`](diffhunk://#diff-93bc202766315b6269beef308a6ad30ed3e86938ddbfa31b49e030f2263695f1R18-R19):
Added a new source and destination for the `release.yml` workflow to
ensure it is included in the sync process.
*
[`.github/workflows/release.yml`](diffhunk://#diff-87db21a973eed4fef5f32b267aa60fcee5cbdf03c67fafdc2a9b553bb0b15f34R1-R32):
Created a new workflow named `release` that triggers on closed pull
requests. This workflow includes steps for setting up the environment,
checking out the repository, setting environment variables, and
publishing a release.