chore(*): move `FUNDING.yml` files to `.github` repository (#389)
This pull request removes the funding configuration for the `lumirlumir`
project and updates the synchronization configuration to stop syncing
the removed funding file. These changes are focused on cleaning up
unused or outdated configuration files.

Configuration cleanup:

*
[`.github/FUNDING.yml`](diffhunk://#diff-07985fdcade0e64d11482724879a644f07879ba61b8fb6c6119e1b1902b72ae4L1-L2):
Removed the funding configuration for the `lumirlumir` project,
including references to GitHub and Open Collective.
*
[`.github/sync-client.yml`](diffhunk://#diff-93bc202766315b6269beef308a6ad30ed3e86938ddbfa31b49e030f2263695f1L29-L30):
Updated the synchronization configuration by removing the entry for
syncing the `FUNDING.yml` file.