chore(*): replace `shx` with native Node.js file operations in scripts (#397)
This pull request updates the `examples/git-clang-format/package.json`
file to replace the use of `shx` with a native Node.js solution and
removes the `shx` dependency from the project's `package.json`.

Dependency updates:

*
[`package.json`](diffhunk://#diff-7ae45ad102eab3b6d7e7896acd08c427a9b25b346470d7bc6507b6481575d519L64):
Removed the `shx` dependency, as it is no longer needed after switching
to a native Node.js approach.

Script updates:

*
[`examples/git-clang-format/package.json`](diffhunk://#diff-4b5616cba814f4c7ca4ea49c77e8456efcc7931852306a3d899597507ad137b6L6-R6):
Replaced the `shx` command in the `add-a-space-to-line-9-of-main-c-file`
script with a Node.js `fs.copyFileSync` operation for copying files.