fix(*): drop `shx` and replace it with Node.js native script (#402)
This pull request refactors the handling of file permissions across
three packages (`clang-format-git-python`, `clang-format-git`, and
`clang-format-node`) by replacing the use of the `shx` dependency with
custom `chmod.js` scripts. These scripts ensure that the necessary files
and binaries have the correct execution permissions. Additionally, the
`postinstall` and `postbuild` scripts are updated to use these new
`chmod.js` files, and the `shx` dependency is removed.

### File Permission Handling Updates:
*
[`packages/clang-format-git-python/chmod.js`](diffhunk://#diff-27b69f71118a60e9a2ad54a11e79e10b8485e6080af45ec037ee193fd598c77cR1-R26):
Added a script to set execution permissions for the `git-clang-format`
script using `chmodSync`.
*
[`packages/clang-format-git/chmod.js`](diffhunk://#diff-337db8b3485bbb276d0c1b16655b094b11cf7605b7adc8e52a4b3f06d0e9edabR1-R33):
Added a script to set execution permissions for binaries in the `bin`
directory, filtering for files containing `git-clang-format`.
*
[`packages/clang-format-node/chmod.js`](diffhunk://#diff-36fbb1ea12298f23d64d479606f6b239ed8a2ca3373684dbd49b5037f16bbfe6R1-R36):
Added a script to set execution permissions for binaries in the `bin`
directory, filtering for files containing `clang-format`.

### Package Configuration Updates:
*
[`packages/clang-format-git-python/package.json`](diffhunk://#diff-fb0eb5dba5b2656b3fdea4322c2ed2a691973f2e9ea536ce20b36c347ab1267eR9):
Added `chmod.js` to the `files` array, replaced `npm run chmod` with
`node chmod.js` in the `postinstall` script, and removed the `shx`
dependency.
[[1]](diffhunk://#diff-fb0eb5dba5b2656b3fdea4322c2ed2a691973f2e9ea536ce20b36c347ab1267eR9)
[[2]](diffhunk://#diff-fb0eb5dba5b2656b3fdea4322c2ed2a691973f2e9ea536ce20b36c347ab1267eL52-R61)
*
[`packages/clang-format-git/package.json`](diffhunk://#diff-d85dcfb0934f579cff896723ab5eeb9a52668d2bb322338dddd93bc1241aa4f1R9):
Added `chmod.js` to the `files` array, replaced `npm run chmod` with
`node chmod.js` in the `postinstall` script, and removed the `shx`
dependency.
[[1]](diffhunk://#diff-d85dcfb0934f579cff896723ab5eeb9a52668d2bb322338dddd93bc1241aa4f1R9)
[[2]](diffhunk://#diff-d85dcfb0934f579cff896723ab5eeb9a52668d2bb322338dddd93bc1241aa4f1L51-R60)
*
[`packages/clang-format-node/package.json`](diffhunk://#diff-6e04d643c57b851ab02ad78e41531387d2cc01caf6c4cfbaa536aadd37346a30R9):
Added `chmod.js` to the `files` array, replaced `npm run chmod` with
`node chmod.js` in the `postinstall` script, and removed the `shx`
dependency.
[[1]](diffhunk://#diff-6e04d643c57b851ab02ad78e41531387d2cc01caf6c4cfbaa536aadd37346a30R9)
[[2]](diffhunk://#diff-6e04d643c57b851ab02ad78e41531387d2cc01caf6c4cfbaa536aadd37346a30L50-L60)