ci(*): switch to trusted publishing (#483)
This pull request makes a small update to the npm publishing workflow.
The most notable change is the addition of a step to update npm to the
latest version before installing dependencies and publishing the
package.

Workflow improvements:

* Added a step to update npm to the latest version using `npm install -g
npm@latest` before running `npm ci` and publishing.
(.github/workflows/publish.yml)